cleaned code and rewrote readme
