+++ .gitignore
@@ -1,1 +1,3 @@
 testfile2.txt
+testfile2OUT.txt
+testfile2.filesqueeze
