+++ Manifest.txt
@@ -1,0 +1,1 @@
+Main-Class: FileSqueeze
